+++ test/e2e/workspace/run/stylelint.config.js
@@ -1,0 +1,10 @@
+'use strict';
+
+/** @type {import('stylelint').Config} */
+const config = {
+	rules: {
+		'color-hex-length': 'long',
+	},
+};
+
+module.exports = config;